requirements.txt
anthropic
chromadb==0.4.20
cohere
faster-whisper
langchain
langchain-experimental
mistletoe
openai
pyTelegramBotAPI
python-dotenv
